Refactor submodule in `downloader`: `geofabrik` → `_geofabrik`
